bot/requirements.txt
discord.py>=2.2.0
google-generative-ai>=0.10.0
fastapi>=0.90.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0
aiosqlite>=0.18.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
yt-dlp>=2024.12.0
PyNaCl>=1.5.0
python-socketio[asyncio_client]>=5.8.0
psutil>=5.9.0
typing_extensions
